Подготавливаем файлы к ДЗ
